+++ src/main/java/com/banner/bannerApplication/repositories/UserRepository.java
@@ -1,7 +1,8 @@
-package com.banner.bannerApplication;
+package com.banner.bannerApplication.repositories;
 
 import java.util.List;
 
+import com.banner.bannerApplication.entities.User;
 import org.springframework.data.repository.PagingAndSortingRepository;
 import org.springframework.data.repository.query.Param;
 import org.springframework.data.rest.core.annotation.RepositoryRestResource;
